refactor: extract wall aerial stats accumulator
